crash: use new gdb.Target to create standalone target
When I rebased crash-python-gdb to an 8.3 prerelease, I found that
targets have been converted to C++.  That necessitated a rewrite of
much of the target code, and I cleaned up some rough edges.

With the new target, we load the vmcore using a simple
'target kdumpfile /path/to/vmcore' command that can be used entirely
outside of the crash semantic code.  This means we can debug the
target more easily and use it in the testing code without having to
parse everything for every test.

This commit converts crash to use the new target but doesn't exploit it
for testing yet.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>