kdump.target: don't specify offset for .data..percpu
The "right" thing to do is for .data..percpu to be loaded at offset 0.
Unfortunately, that only works when debuginfo is embedded in the binary.
When separate debuginfo is used, section offsets can't be specified and
gdb interprets an offset of 0 to mean "immediately after the preceding
section."  In order to make the rest of the percpu code sane, we'll
let gdb make the same assumption when embedded debuginfo is used.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>